+++ packages/ovee/src/core/App.ts
@@ -1,22 +1,23 @@
 import { FRAMEWORK_NAME } from 'src/constants';
 import { OveeComponent } from 'src/core/types';
-import { Callback, EventDelegate, EventDesc } from 'src/dom';
+import { Callback, EventDelegate, EventDesc, ListenerOptions, TargetOptions } from 'src/dom';
 import { ComponentError } from 'src/errors';
 import {
+	AnyObject,
 	attachMutationObserver,
-	Class,
 	ClassConstructor,
 	Dictionary,
+	isString,
 	isValidNode,
 } from 'src/utils';
 
-import Component, { ComponentOptions } from './Component';
-import Module, { ModuleOptions } from './Module';
+import Component, { ComponentClass, ComponentOptions } from './Component';
+import Module, { ModuleClass, ModuleOptions } from './Module';
 
 export interface AppConstructorParams {
 	config?: Partial<AppConfig>;
-	components?: RegisterPayload<Class<Component, typeof Component>, ComponentOptions>;
-	modules?: RegisterPayload<ClassConstructor<Module>, ModuleOptions>;
+	components?: RegisterPayload<ComponentClass, ComponentOptions>;
+	modules?: RegisterPayload<ModuleClass, ModuleOptions>;
 }
 
 export interface AppConfig {
@@ -28,8 +29,8 @@ export interface AppConfig {
 
 type RegisterPayload<C, O> = (C | [C, O])[];
 
-export interface ComponentStorage<C extends Component> {
-	ComponentClass: Class<C, typeof Component>;
+export interface ComponentStorage {
+	ComponentClass: ComponentClass;
 	options: ComponentOptions;
 }
 
@@ -42,7 +43,7 @@ const defaultConfig: AppConfig = {
 
 export default class App {
 	readonly modules: Dictionary<Module> = {};
-	readonly components: Dictionary<ComponentStorage<Component>> = {};
+	readonly components: Dictionary<ComponentStorage> = {};
 	readonly $eventDelegate!: EventDelegate<any>;
 	initialized = false;
 	rootElement?: Element;
@@ -106,7 +107,7 @@ export default class App {
 		return { ...this.config };
 	}
 
-	registerModules(modules: RegisterPayload<ClassConstructor<Module>, ModuleOptions>): this {
+	registerModules(modules: RegisterPayload<ModuleClass, ModuleOptions>): this {
 		modules.forEach(module => {
 			if (Array.isArray(module)) {
 				this.use(module[0], module[1]);
@@ -118,7 +119,7 @@ export default class App {
 		return this;
 	}
 
-	use<M extends Module, Opt = M extends Module<infer O> ? O : Record<string, any>>(
+	use<M extends Module, Opt = M extends Module<infer O> ? O : AnyObject>(
 		ModuleClass: ClassConstructor<M>,
 		options?: Opt
 	): M {
@@ -145,17 +146,27 @@ export default class App {
 		Object.values(this.modules).forEach(module => module.destroy());
 	}
 
-	getModule(name: string): Module {
-		if (!this.modules[name]) {
-			throw new Error(`Module "${name}" is not registered`);
+	getModule<M extends ModuleClass = ModuleClass>(module: M | string): InstanceType<M> {
+		let name: string;
+
+		if (isString(module)) {
+			name = module;
+		} else {
+			if (!('getName' in module) && typeof module.getName !== 'function') {
+				throw new Error(`Passed classed is not an instance of the 'Module' class`);
+			}
+
+			name = module.getName();
 		}
 
-		return this.modules[name];
+		if (!this.modules[name]) {
+			throw new Error(`Module '${name}' is not registered`);
+		}
+
+		return this.modules[name] as InstanceType<M>;
 	}
 
-	registerComponents(
-		components: RegisterPayload<Class<Component, typeof Component>, ComponentOptions>
-	): this {
+	registerComponents(components: RegisterPayload<ComponentClass, ComponentOptions>): this {
 		components.forEach(component => {
 			if (Array.isArray(component)) {
 				this.registerComponent(component[0], component[1]);
@@ -167,8 +178,8 @@ export default class App {
 		return this;
 	}
 
-	registerComponent<C extends Component>(
-		ComponentClass: Class<C, typeof Component>,
+	registerComponent<C extends ComponentClass>(
+		ComponentClass: C,
 		options: ComponentOptions = {}
 	): this {
 		if (!(ComponentClass.prototype instanceof Component)) {
@@ -240,15 +251,15 @@ export default class App {
 		);
 	}
 
-	makeComponent<C extends Component>(
+	makeComponent<C extends ComponentClass>(
 		el: Element & OveeComponent,
-		ComponentClass: Class<C, typeof Component>,
+		ComponentClass: C,
 		options: ComponentOptions = {}
-	): C {
+	): InstanceType<C> {
 		// @todo handle multi-component elements
 		if (!el._oveeComponentInstance) {
 			el.classList.add(`${this.config.namespace}-component`);
-			el._oveeComponentInstance = new ComponentClass(el, this, options) as C;
+			el._oveeComponentInstance = new ComponentClass(el, this, options);
 		} else if (!(el._oveeComponentInstance instanceof ComponentClass)) {
 			throw new ComponentError(
 				'Component instance has already been initialized for this element',
@@ -257,7 +268,7 @@ export default class App {
 			);
 		}
 
-		return el._oveeComponentInstance as C;
+		return el._oveeComponentInstance as InstanceType<C>;
 	}
 
 	destroyComponent(el: Element & OveeComponent): void {
@@ -281,20 +292,12 @@ Make sure to turn on production mode when deploying for production.`
 		}
 	}
 
-	$on(events: string, callback: Callback<any>): this;
-	$on(events: string, target: Element, callback: Callback<any>): this;
-	$on(events: string, target: any, callback?: any): this {
-		this.$eventDelegate.on(events, target, callback);
-
-		return this;
+	$on(events: string, callback: Callback<any>, options?: ListenerOptions): () => void {
+		return this.$eventDelegate.on(events, callback, options);
 	}
 
-	$off(events: string, callback: Callback<any>): this;
-	$off(events: string, target: Element, callback: Callback<any>): this;
-	$off(events: string, target: any, callback?: any): this {
-		this.$eventDelegate.off(events, target, callback);
-
-		return this;
+	$off(events: string, callback: Callback<any>, options?: TargetOptions): void {
+		return this.$eventDelegate.off(events, callback, options);
 	}
 
 	$emit<D = any>(eventDesc: EventDesc, detail?: D): void {