+++ packages/ovee/src/dom/EventDelegate.ts
@@ -1,145 +1,136 @@
-import Component from 'src/core/Component';
+import { Logger } from 'src/errors';
+import { isString, omit } from 'src/utils';
 
 import { emitEvent } from './emitEvent';
 
-interface Listener<T extends Component> {
-	event: string;
-	target: Element;
-	selector?: string;
-	callback: Callback<T>;
-	handler: (...args: any[]) => void;
-}
-
-interface CastedParams<C extends Component> {
-	events: string;
-	target: Element;
-	selector?: string;
-	callback: Callback<C>;
-}
-
-export interface Callback<T extends Component> extends Function {
+export interface Callback<T> extends Function {
 	(this: T, ...args: any[]): void;
 }
 
+type Target = string | EventTarget | EventTarget[];
+
+export interface TargetOptions {
+	target?: Target;
+	root?: true;
+	multiple?: boolean;
+}
+
+export interface ListenerOptions extends AddEventListenerOptions, TargetOptions {}
+
 export type EventDesc = string | Event;
 
-export class EventDelegate<Context extends Component> {
-	listeners: Listener<Context>[] = [];
-	targetElement: Element;
-	context: Context;
+interface Listener {
+	cleared: boolean;
+	events: string;
+	target: Target;
+	callback: Callback<any>;
+	removeListeners: Array<() => void>;
+}
 
-	constructor(targetElement: Element, context: Context) {
-		this.targetElement = targetElement;
-		this.context = context;
-	}
+const logger = new Logger('EventDelegate');
 
-	on(events: string, callback: Callback<Context>): void;
-	on(events: string, selector: string, callback: Callback<Context>): void;
-	on(events: string, target: Element, callback: Callback<Context>): void;
-	on(events: string, target: Element, selector: string, callback: Callback<Context>): void;
-	on(
-		_events: string,
-		_target: Callback<Context> | string | Element,
-		_selector?: Callback<Context> | string,
-		_callback?: Callback<Context>
-	): void {
-		const { events, target, selector, callback } = this.getCastedParams(
-			_events,
-			_target,
-			_selector,
-			_callback
-		);
+export class EventDelegate<Context = any> {
+	listeners: Listener[] = [];
+
+	constructor(public targetElement: Element, public context: Context) {}
+
+	on(events: string, callback: Callback<Context>, options?: ListenerOptions): () => void {
+		const listenerOptions = options
+			? (omit(options ?? {}, ['root', 'target', 'multiple']) as AddEventListenerOptions)
+			: undefined;
+		const { targetOption, target } = this.getTarget(options);
+		const removeListeners: Array<() => void> = [];
 
 		events.split(' ').forEach(event => {
-			let handler;
-			let capture;
+			const handler = (...args: any[]) => callback.apply(this.context, args);
+			const targets = Array.isArray(target) ? target : [target];
 
-			if (selector !== undefined) {
-				handler = (...args: any[]) => {
-					const [e] = args;
+			targets.forEach(t => {
+				t.addEventListener(event, handler, listenerOptions);
+			});
 
-					if (e.target && (e.target.matches(selector) || e.target.closest(selector))) {
-						callback.apply(this.context, args);
-					}
-				};
-				capture = true;
-			} else {
-				handler = (...args: any[]) => {
-					callback.apply(this.context, args);
-				};
-				capture = false;
-			}
-
-			target.addEventListener(event, handler, { capture });
-
-			this.listeners.push({
-				event,
-				target,
-				selector,
-				callback,
-				handler,
+			removeListeners.push(() => {
+				targets.forEach(t => t.removeEventListener(event, handler));
 			});
 		});
+
+		const listener: Listener = {
+			cleared: false,
+			events,
+			callback,
+			removeListeners,
+			target: targetOption,
+		};
+
+		this.listeners.push(listener);
+
+		return () => this.clearListener(listener);
 	}
 
-	off(events: string, callback: Callback<Context>): void;
-	off(events: string, selector: string, callback: Callback<Context>): void;
-	off(events: string, target: Element, callback: Callback<Context>): void;
-	off(events: string, target: Element, selector: string, callback: Callback<Context>): void;
-	off(
-		_events: string,
-		_target: Callback<Context> | string | Element,
-		_selector?: Callback<Context> | string,
-		_callback?: Callback<Context>
-	): void {
-		const { events, target, selector, callback } = this.getCastedParams(
-			_events,
-			_target,
-			_selector,
-			_callback
+	off(events: string, callback: Callback<Context>, options?: TargetOptions): void {
+		const { targetOption } = this.getTarget(options);
+		const listener = this.listeners.find(
+			l =>
+				l.events === events &&
+				l.callback === callback &&
+				this.areTargetsTheSame(l.target, targetOption)
 		);
 
-		events.split(' ').forEach(event => {
-			this.listeners.forEach((item, index) => {
-				if (
-					item.event === event &&
-					item.target === target &&
-					item.selector === selector &&
-					item.callback === callback
-				) {
-					target.removeEventListener(event, item.handler);
-					this.listeners.splice(index, 1);
-				}
-			});
-		});
+		if (listener) this.clearListener(listener);
 	}
 
-	private getCastedParams(
-		events: string,
-		target: Callback<Context> | string | Element,
-		selector?: Callback<Context> | string,
-		callback?: Callback<Context>
-	) {
-		const out: Partial<CastedParams<Context>> = { events, callback };
+	private areTargetsTheSame(targetA: Target, targetB: Target): boolean {
+		if (Array.isArray(targetA)) {
+			if (!Array.isArray(targetB)) return false;
 
-		if (callback === undefined) {
-			if (selector === undefined) {
-				out.callback = target as Callback<Context>;
-				out.target = this.targetElement;
-			} else if (typeof target === 'string') {
-				out.callback = selector as Callback<Context>;
-				out.selector = target;
-				out.target = this.targetElement;
-			} else {
-				out.callback = selector as Callback<Context>;
-				out.target = target as Element;
-			}
-		} else {
-			out.target = target as Element;
-			out.selector = selector as string;
+			return targetA.every((v, i) => targetB[i] === v);
 		}
 
-		return out as CastedParams<Context>;
+		return targetA === targetB;
+	}
+
+	private getTarget(options?: TargetOptions): {
+		target: EventTarget | EventTarget[];
+		targetOption: Target;
+	} {
+		const targetOption = options?.target ?? this.targetElement;
+		let target: EventTarget | EventTarget[];
+
+		if (isString(targetOption)) {
+			const isAbsolute = options?.root;
+			const multipleTargets = options?.multiple;
+			const selectorBase = isAbsolute ? document : this.targetElement;
+			const newTarget = multipleTargets
+				? Array.from(selectorBase.querySelectorAll(targetOption))
+				: selectorBase.querySelector(targetOption);
+
+			if (!newTarget || (Array.isArray(newTarget) && !newTarget.length)) {
+				const errorMessage = `Could not find element${
+					multipleTargets ? 's' : ''
+				} with selector '${targetOption}' ${
+					isAbsolute ? 'in document' : 'relatively to current element'
+				}`;
+
+				throw Error(logger.getMessage(errorMessage));
+			}
+
+			target = newTarget;
+		} else {
+			target = targetOption;
+		}
+
+		return { target, targetOption };
+	}
+
+	private clearListener(listener: Listener) {
+		if (listener.cleared) return;
+
+		listener.removeListeners.forEach(cb => cb());
+		listener.cleared = true;
+
+		const idx = this.listeners.findIndex(l => l === listener);
+
+		if (idx >= 0) this.listeners.splice(idx, 1);
 	}
 
 	emit<D = any>(eventDesc: EventDesc, detail?: D): void {
@@ -148,7 +139,8 @@ export class EventDelegate<Context extends Component> {
 
 	destroy(): void {
 		this.listeners = this.listeners.filter(item => {
-			item.target.removeEventListener(item.event, item.handler);
+			item.removeListeners.forEach(cb => cb());
+
 			return false;
 		});
 	}